requirements.txt
fastapi
uvicorn
httpx
openai
serpapi
alpha_vantage
yfinance
matplotlib
python-dotenv
google-search-results
mcp[cli]
